feat: implement global PageId and atomic write mechanism in DiskManager

+++ crates/wackdb_common/src/constants.rs
@@ -2,7 +2,7 @@ use crate::types::{FrameId, Lsn, PageId, SlotId, TransactionId};
 
 pub const PAGE_SIZE: usize = 8192;
 
-pub const INVALID_PAGE_ID: PageId = PageId(u32::MAX);
+pub const INVALID_PAGE_ID: PageId = PageId(u64::MAX);
 
 pub const INVALID_FRAME_ID: FrameId = FrameId(u32::MAX);
